Renamed models directory to generalModels and renamed mainMenu to discussionMainMenu
