+++ src/generalModels/Course.ts
