+++ src/interaction/command/interface.Command.ts
@@ -2,12 +2,9 @@ import { CommandInteraction, ChatInputApplicationCommandData, Client } from "dis
 
 /**
  * @interface Command
- * 
  * @description Interface that extends discords default ChatInputApplicationCommandData
  * useful because we can specify a run property so that we can declare command behavior with the definition of the command
- * 
  * @property {Function} run - the function to be run when the command is called
- * 
  */
 export interface Command extends ChatInputApplicationCommandData {
     run: (client: Client, interaction: CommandInteraction) => void;
